requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
Markdown==3.5.2
multidict==6.0.5
outcome==1.3.0.post0
pillow==10.2.0
pycparser==2.21
PySocks==1.7.1
pytz==2024.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
